requirements.txt
transformers==4.45.1
accelerate>=0.26.0
gradio
sqlparser
